config.fish: changed starship to be only written, when not existing

+++ config.fish
@@ -6,8 +6,7 @@ if status is-interactive
     fish_config theme choose --color-theme=dark "Rosé Pine Moon Auto"
 
     # starship theme
-    if command -q starship && test -e ~/.config/fish/plugins/rose-pine-starship/rose-pine-moon.toml
-        rm ~/.config/starship.toml >/dev/null
+    if command -q starship && test -e ~/.config/fish/plugins/rose-pine-starship/rose-pine-moon.toml && ! test -e ~/.config/starship.toml
         ln -s ~/.config/fish/plugins/rose-pine-starship/rose-pine-moon.toml ~/.config/starship.toml
     end
